+++ R/action.R
@@ -15,26 +15,51 @@ add_action_impl <- function(x, action, name, ..., call = caller_env()) {
 
 add_action_impl.action_pre <- function(x, action, name, ..., call = caller_env()) {
   check_singleton(x$pre$actions, name, call = call)
-  x$pre <- add_action_to_stage(x$pre, action, name)
+  x$pre <- add_action_to_stage(x$pre, action, name, order_stage_pre())
   x
 }
 
 add_action_impl.action_fit <- function(x, action, name, ..., call = caller_env()) {
   check_singleton(x$fit$actions, name, call = call)
-  x$fit <- add_action_to_stage(x$fit, action, name)
+  x$fit <- add_action_to_stage(x$fit, action, name, order_stage_fit())
   x
 }
 
 add_action_impl.action_post <- function(x, action, name, ..., call = caller_env()) {
   check_singleton(x$post$actions, name, call = call)
-  x$post <- add_action_to_stage(x$post, action, name)
+  x$post <- add_action_to_stage(x$post, action, name, order_stage_post())
   x
 }
 
 # ------------------------------------------------------------------------------
 
-add_action_to_stage <- function(stage, action, name) {
-  stage$actions <- c(stage$actions, list2(!!name := action))
+order_stage_pre <- function() {
+  # Case weights must come before preprocessor
+  c(
+    c("case_weights"),
+    c("formula", "recipe", "variables")
+  )
+}
+
+order_stage_fit <- function() {
+  "model"
+}
+
+order_stage_post <- function() {
+  character()
+}
+
+# ------------------------------------------------------------------------------
+
+add_action_to_stage <- function(stage, action, name, order) {
+  actions <- c(stage$actions, list2(!!name := action))
+
+  # Apply required ordering for this stage
+  order <- intersect(order, names(actions))
+  actions <- actions[order]
+
+  stage$actions <- actions
+
   stage
 }